pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Compare two configuration files"
authors = ["Aigars Sinevics <aigarzs@gmail.com>"]
readme = "README.md"
packages = [{ include = "scripts" },
	{ include =  "gendiff" },
	{ include = "tests" }
	]
 
[tool.poetry.dependencies]
python = "^3.10"
#lcov = "^1.15.5a0"
pyyaml = "^6.0.1"



[tool.poetry.scripts]
gendiff = "scripts.gendiff:main"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
#clover = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
